+++ HQServer/HQServer.Domain/Entities/OutletTransactionDetail.cs
@@ -16,6 +16,6 @@ namespace HQServer.Domain.Entities
         [Key, Column(Order = 2)]
         public string barcode { get; set; }
         public int unitSold { get; set; }
-        public double cost { get; set; }
+        public decimal cost { get; set; }
     }
 }
